Ping, netfps and timeout refinements

+++ client.h
@@ -657,8 +657,8 @@ typedef struct usercmd_s
 	int		cursor_entitynumber;
 
 	double time; // time the move is executed for (cl_movement: clienttime, non-cl_movement: receivetime)
-	double receivetime; // time the move was received at
-	double clienttime; // time to which server state the move corresponds to
+	float receivetime; // time the move was received at
+	float clienttime; // time to which server state the move corresponds to (used to calculate ping)
 	int msec; // for predicted moves
 	int buttons;
 	int impulse;
